Refactored to have collection and review state in dashboard component
